requirements.txt
openai
python-dotenv
PyMuPDF
Pillow
numpy
langchain_openai
qdrant_client